better placeholder selection (like with the overlays before), \begin{} magic for placeholder mirror example
git-svn-id: svn://svn.code.sf.net/p/texstudio/code/trunk@324 92cce386-e96d-4106-809f-596c673a3fd8

+++ qcodeedit/lib/qeditorinputbinding.cpp
@@ -161,13 +161,13 @@ bool QEditorInputBinding::keyPressEvent(QKeyEvent *event, QEditor *editor)
 		int idx = m_index.at(i);
 		const QKeySequence& ks = m_keys.at(i);
 		
-		if ( idx < ks.count() )
+		if ( idx < (int) ks.count() )
 		{
 			if ( ks[idx] == event->key() )
 			{
 				++idx;
 				
-				if ( idx == ks.count() )
+				if ( idx == (int)ks.count() )
 				{
 					qDebug("match");
